Create /etc/init/softether.conf for upstart.
Create a softether.upstart file in the Debian overlay for systems that use
the upstart init subsystem.  This lets the operator control the SoftEther
daemon with commands like:

	# start softether
	# stop softether
	# status softether
	# restart softether

Also modify the softether.init file to exit early if the /etc/init/softether
file is installed.  Although the dh_installinit debhelper prefers an upstart
configuration for automatic control, it will also install the sysv script.