AppVeyor: remove MSVC 2008 build
The purpose of the configuration was to build installers and continually verify that MSVC 2008 could still build the project.

MSVC 2008 was a requirement because we wanted to maintain support for very old versions of Windows (9x).

During the past few years we encountered many annoying limitations due to that requirement, such as the lack of support for designated initializers.

We ended up deciding to only keep support for operating systems older than 7 in the stable repository.

As for the installers, we now build them on Azure Pipelines (currently with MSVC 2019).